remove logging of query parameters in getAll function; update quantity type to DECIMAL(10, 3) in sales_products migration; adjust seed values for quantity in sales products seeder

+++ backend/src/database/migrations/05-create-sales-products.ts
@@ -34,7 +34,7 @@ export default {
         onUpdate: 'CASCADE',
       },
       quantity: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(10, 3),
         allowNull: false,
       },
     });
